Refactor: extract DeviceArgs + KernelArgsHelper to src/common/platform/onboard/host
Wires up a new shared module `src/common/platform/onboard/host/` for code
that's line-identical between a2a3 and a5 onboard `libhost_runtime.so`,
and moves the first two structs into it:

- `DeviceArgs` (per-task AICPU device-args, layout identical on both archs).
- `KernelArgsHelper` (6 common methods: init/finalize × {device_args,
  runtime_args, device_kernel_args}).

a2a3-only extensions (which can't live on the common struct because a5's
`KernelArgs` has no matching field):

- `init_ffts_base_addr` becomes a free function
  `kernel_args_init_ffts_base_addr(KernelArgsHelper &)` in a2a3's
  `device_runner.h`. One callsite updated.

Build / link:

- Both arch `CMakeLists.txt` add `src/common/platform/onboard/host` to
  the include path and pull in `device_runner_helpers.cpp`.
- The common header `#include "common/kernel_args.h"` resolves to the
  arch-specific `KernelArgs` layout via each arch's existing include
  path — no change in arch ABI.

Net delta: -185 LoC (-400 / +215 across 6 files). ABI / Python surface
unchanged. Sim onboard untouched.

Local validation (a2a3 onboard, device 1):
- vector_example: PASS
- aicore_op_timeout: PASS

See `.docs/ONBOARD_HOST_COMMON_REFACTOR.md` for the full migration plan
(this is PR 1 + PR 2 of the planned series — scaffolding + first
extraction collapsed into one PR after PR #877 removed `AicpuSoInfo`).
