+++ src/a2a3/platform/onboard/host/device_runner.h
@@ -44,6 +44,7 @@
 #include "common/platform_config.h"
 #include "common/unified_log.h"
 #include "device_arena.h"
+#include "device_runner_helpers.h"  // common DeviceArgs + KernelArgsHelper
 #include "host/function_cache.h"
 #include "host/memory_allocator.h"
 #include "host/l2_perf_collector.h"
@@ -54,106 +55,14 @@
 #include "runtime.h"
 
 /**
- * DeviceArgs structure for AICPU device arguments.
+ * a2a3-only `KernelArgsHelper` extension: retrieve the FFTS base address via
+ * `rtGetC2cCtrlAddr` and store it in the wrapped `KernelArgs`. a5's
+ * `KernelArgs` has no `ffts_base_addr` field, so this helper lives in the
+ * arch-specific header rather than on the common `KernelArgsHelper` struct.
  *
- * Layout offsets are still nominally fixed by libaicpu_extend_kernels.so for
- * aicpu_so_bin / aicpu_so_len (at offsets 96 / 104), but per-task AICPU
- * launches go through rtsLaunchCpuKernel against the cached rtFuncHandle on
- * LoadAicpuOp — none of our code reads these fields. The fields are kept
- * (zero-initialized, never assigned) so the H2D struct layout matches the
- * historical contract on both archs; an earlier "the H2D allocation pointed
- * to by aicpu_so_bin is load-bearing on a5 onboard" finding no longer
- * reproduces against current HEAD (post #864/#870), so the runner-side
- * AicpuSoInfo allocation was removed.
+ * @return 0 on success, error code on failure.
  */
-struct DeviceArgs {
-    uint64_t unused[12] = {0};
-    uint64_t aicpu_so_bin{0};
-    uint64_t aicpu_so_len{0};
-};
-
-/**
- * Helper class for managing KernelArgs with device memory
- *
- * This class wraps KernelArgs and provides host-side initialization methods
- * for allocating device memory and copying data to the device. It separates
- * the concerns of device memory management (host-only) from the structure
- * layout (shared with kernels).
- *
- * The helper provides implicit conversion to KernelArgs* for seamless use
- * with runtime APIs.
- */
-struct KernelArgsHelper {
-    KernelArgs args;
-    MemoryAllocator *allocator_{nullptr};
-    KernelArgs *device_k_args_{nullptr};  // Device copy of KernelArgs for AICore
-
-    /**
-     * Initialize device arguments by allocating device memory and copying data
-     *
-     * @param host_device_args  Host-side device arguments to copy
-     * @param allocator       Memory allocator to use
-     * @return 0 on success, error code on failure
-     */
-    int init_device_args(const DeviceArgs &host_device_args, MemoryAllocator &allocator);
-
-    /**
-     * Free device memory allocated for device arguments
-     *
-     * @return 0 on success, error code on failure
-     */
-    int finalize_device_args();
-
-    /**
-     * Initialize runtime arguments by allocating device memory and copying data
-     *
-     * @param host_runtime  Host-side runtime to copy to device
-     * @param allocator  Memory allocator to use
-     * @return 0 on success, error code on failure
-     */
-    int init_runtime_args(const Runtime &host_runtime, MemoryAllocator &allocator);
-
-    /**
-     * Free device memory allocated for runtime arguments
-     *
-     * @return 0 on success, error code on failure
-     */
-    int finalize_runtime_args();
-
-    /**
-     * Retrieve FFTS base address via rtGetC2cCtrlAddr and store in KernelArgs
-     *
-     * @return 0 on success, error code on failure
-     */
-    int init_ffts_base_addr();
-
-    /**
-     * Copy KernelArgs to device memory for AICore kernel parameter passing
-     *
-     * Must be called after init_runtime_args and init_ffts_base_addr.
-     *
-     * @param allocator  Memory allocator to use
-     * @return 0 on success, error code on failure
-     */
-    int init_device_kernel_args(MemoryAllocator &allocator);
-
-    /**
-     * Free device memory allocated for KernelArgs copy
-     *
-     * @return 0 on success, error code on failure
-     */
-    int finalize_device_kernel_args();
-
-    /**
-     * Implicit conversion operators for seamless use with runtime APIs
-     *
-     * These operators allow KernelArgsHelper to be used wherever KernelArgs*
-     * is expected, enabling transparent device memory management while
-     * maintaining API compatibility.
-     */
-    operator KernelArgs *() { return &args; }
-    KernelArgs *operator&() { return &args; }
-};
+int kernel_args_init_ffts_base_addr(KernelArgsHelper &helper);
 
 /**
  * Device runner for kernel execution
